[Kernel] Relocate dspark kernel subpackage into kernels.ops.speculative (RFC #29630)
srt/speculative/dspark_components/kernels/ was a dedicated kernel subpackage
(dspark accept / verify_window / attn_metadata / draft_model / schedule +
dispatch) sitting in a folder literally named `kernels/`. Move it wholesale to
sglang.kernels.ops.speculative.dspark and rewrite all callers. The subpackage
has no back-import to its dspark_components parent, so the relocation is clean.

Validated: find_spec resolves, namespace import stays metadata-only (CLEAN),
py_compile + pre-commit clean.
