+++ python/sglang/srt/speculative/dspark_components/dspark_verify.py
@@ -5,6 +5,23 @@ from typing import Optional
 import msgspec
 import torch
 
+from sglang.kernels.ops.speculative.dspark.dspark_accept import (
+    AcceptGreedy,
+    AcceptSampling,
+    FinalizeAcceptLens,
+    SelectMixedAccept,
+    SoftmaxTemp,
+    accept_greedy_triton,
+    finalize_accept_lens_triton,
+)
+from sglang.kernels.ops.speculative.dspark.dspark_verify_window import (
+    BuildCommitInjectLayout,
+    BuildOutTokens,
+    BuildRaggedVerifyWindow,
+    RaggedVerifyWindow,
+    ScatterCompactToStrided,
+    scatter_compact_to_strided_into,
+)
 from sglang.srt.layers.logits_processor import LogitsProcessorOutput
 from sglang.srt.managers.schedule_batch import ScheduleBatch
 from sglang.srt.model_executor.forward_batch_info import CaptureHiddenMode, ForwardMode
@@ -18,23 +35,6 @@ from sglang.srt.speculative.dspark_components.dspark_kv_inject import (
 from sglang.srt.speculative.dspark_components.dspark_planner import (
     VerifyWindow,
     apply_logits_adjustments_strided,
-)
-from sglang.srt.speculative.dspark_components.kernels.dspark_accept import (
-    AcceptGreedy,
-    AcceptSampling,
-    FinalizeAcceptLens,
-    SelectMixedAccept,
-    SoftmaxTemp,
-    accept_greedy_triton,
-    finalize_accept_lens_triton,
-)
-from sglang.srt.speculative.dspark_components.kernels.dspark_verify_window import (
-    BuildCommitInjectLayout,
-    BuildOutTokens,
-    BuildRaggedVerifyWindow,
-    RaggedVerifyWindow,
-    ScatterCompactToStrided,
-    scatter_compact_to_strided_into,
 )
 from sglang.srt.speculative.ragged_verify import RaggedVerifyLayout
