+++ test/registered/spec/dspark/test_dspark_kernel_parity.py
@@ -16,15 +16,15 @@ from sglang.kernels.ops.attention import (
     dsv4_attn_metadata_kernels as attn_metadata_kernels,
 )
 from sglang.kernels.ops.speculative import ragged_verify_kernels
-from sglang.srt.speculative.dspark_components.dspark_planner import (
-    DSparkScheduleConfig,
-)
-from sglang.srt.speculative.dspark_components.kernels import (
+from sglang.kernels.ops.speculative.dspark import (
     dspark_accept,
     dspark_attn_metadata,
     dspark_draft_model,
     dspark_schedule,
     dspark_verify_window,
+)
+from sglang.srt.speculative.dspark_components.dspark_planner import (
+    DSparkScheduleConfig,
 )
 from sglang.srt.speculative.ragged_verify import RaggedVerifyLayout
 from sglang.test.ci.ci_register import register_cuda_ci
